Consolidate all trimming tests into single comprehensive file
- Merged trimming.rs and backwards_compat.rs into inline_snapshot_trimming.rs
- 9 comprehensive tests covering:
  - Basic trimming functionality (4 tests)
  - Backwards compatibility (5 tests)
- Tests cover all critical scenarios:
  - Old format compatibility
  - Migration path via --accept
  - Force-update aggressive trimming
  - Warning behavior
  - File snapshots
- Single file is easier to maintain and understand
